pyproject.toml
[project]
name = "discord-tools"
version = "0.1.0"
description = "Tools for Discord"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "discord-py>=2.4.0",
    "asyncio>=3.4.3",
]
